IMPL add sleep to slow down emulation speed. Better handling of keyboards inputs

+++ Chip8.h
@@ -26,7 +26,7 @@ public:
     void setKeys();
 
     bool drawFlag;
-    std::map<int, unsigned char> keypadMap;
+    std::map<Uint8, unsigned char> keypadMap;
 
     unsigned short pc; // program counter
